Encode keys as zero-padded for lexicographic sorting
