+++ example/src/utils/BlockProcessor.ts
@@ -1,6 +1,3 @@
-import path from "path";
-import levelup, { LevelUp } from "levelup";
-import leveldown from "leveldown";
 import { ServiceError } from "@grpc/grpc-js";
 import { NoteEncrypted } from "@ironfish/sdk/build/src/primitives/noteEncrypted";
 import {
@@ -9,6 +6,7 @@ import {
   LightBlock,
   LightStreamerClient,
 } from "../../../src/models/lightstreamer";
+import { BlockCache } from "./BlockCache";
 
 function addToMerkleTree(note: NoteEncrypted) {
   return note;
@@ -16,22 +14,28 @@ function addToMerkleTree(note: NoteEncrypted) {
 
 const POLL_INTERVAL = 30 * 1000;
 
-const CACHE_KEYS = {
-  HEAD_SEQUENCE: "HEAD_SEQUENCE",
-};
+/**
+ * @todo:
+ * Reorgs =>
+ *   To determine if re-org happened, when querying new blocks, check that each block's prev block hash
+ *   matches the previous block's block hash. If it does not, walk back until you find a block that matches.
+ * Store transactions =>
+ *   Add simple DB to store transactions so we don't have to start querying from block 1 when restarting or
+ *   importing a new account.
+ * Account balances =>
+ *   Add example of processing notes to deterine account balances.
+ * Add error handling to server if unable to connect to node.
+ */
 
 export class BlockProcessor {
   private client: LightStreamerClient;
   private pollInterval?: NodeJS.Timer;
   private handleStop?: () => void;
   private isProcessingBlocks: boolean = false;
-  private blockCache: LevelUp;
+  private blockCache = new BlockCache();
 
   constructor(client: LightStreamerClient) {
     this.client = client;
-    this.blockCache = levelup(
-      leveldown(path.join(__dirname, "..", "client-block-cache")),
-    );
   }
 
   public async start() {
@@ -76,7 +80,7 @@ export class BlockProcessor {
       throw new Error("Head sequence is undefined");
     }
 
-    const cachedHeadSequence = await this._getHeadSequence();
+    const cachedHeadSequence = await this.blockCache.getHeadSequence();
 
     if (headSequence === cachedHeadSequence) {
       return;
@@ -85,19 +89,6 @@ export class BlockProcessor {
     await this._processBlockRange(cachedHeadSequence + 1, headSequence);
 
     this.isProcessingBlocks = false;
-  }
-
-  private async _getHeadSequence() {
-    try {
-      const headSequence = await this.blockCache.get(CACHE_KEYS.HEAD_SEQUENCE);
-      const asNumber = Number(headSequence);
-      if (isNaN(asNumber)) {
-        throw new Error("Head sequence is not a number");
-      }
-      return asNumber;
-    } catch (_err) {
-      return 0;
-    }
   }
 
   private _getLatestBlock() {
@@ -136,7 +127,7 @@ export class BlockProcessor {
   }
 
   private _processBlock(block: LightBlock) {
-    this._cacheBlock(block);
+    this.blockCache.cacheBlock(block);
 
     for (const transaction of block.transactions) {
       for (const output of transaction.outputs) {
@@ -144,16 +135,5 @@ export class BlockProcessor {
         addToMerkleTree(note);
       }
     }
-  }
-
-  private _cacheBlock(block: LightBlock) {
-    const sequence = block.sequence;
-    console.log(`Caching block ${sequence}`);
-
-    this.blockCache
-      .batch()
-      .put(sequence, block)
-      .put(CACHE_KEYS.HEAD_SEQUENCE, sequence)
-      .write();
   }
 }
